Placed getFlockDetails inside ReadFlockDetails (was inside ReadFlock)
